+++ app/(user)/index.tsx
@@ -1,4 +1,4 @@
-import UserSettingsScreen from '@/src/components/screens/user-settings/UserSettingsScreen';
+import UserSettingsScreen from '@/src/components/screens/UserSettingsScreen';
 
 export default function Index() {
   return <UserSettingsScreen />;
